# go/model_add_shopping_cart_request.go
/*
 * Online Store
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type AddShoppingCartRequest struct {

	ProductId int32 `json:"product_id"`
}