# go/model_delete_shopping_cart_request.go
/*
 * Online Store
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type DeleteShoppingCartRequest struct {

	Id int32 `json:"id"`
}