# go/model_email_checking_request.go
/*
 * Online Store
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type EmailCheckingRequest struct {

	Email string `json:"email"`
}