# go/model_register_user_request.go
/*
 * Online Store
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type RegisterUserRequest struct {

	Name string `json:"name"`

	Email string `json:"email"`

	Password string `json:"password"`
}